request invest list from app component and pass it to down to ModelsTab.
