+++ readme.md
@@ -28,7 +28,7 @@ see `package.json` `scripts` object.
 To run these or other command-line utils of locally installed packages outside the context of the `package.json scripts`, use `npx` (e.g. `npx eslint ...`) as a shortcut to the executeable. 
 
 **To run a single test file:**  
-`set INVEST=C:\\Users\\dmf\\Miniconda3\\envs\\invest-py37\\Scripts\\invest.exe && npx jest -u main.test.js`  (note this is not the path to the test file, rather a pattern for matching)  
+`set INVEST=C:\\Users\\dmf\\Miniconda3\\envs\\invest-py37\\Scripts\\invest.exe && npx jest main.test.js`  (note this is not the path to the test file, rather a pattern for matching)  
 
 To run snippets of code outside the electron runtime, but with the same ECMAscript features and babel configurations, use `node -r @babel/register script.js`. This is useful for development of isolated modules.
